Refactor WorkerActor and JobCoordinatorGrain for improved job handling and telemetry
- Enhanced WorkerActor to manage work queue with prefetching and processing logic.
- Introduced new messages: PrefetchWork, ProcessWork, and HeartbeatTick for better actor communication.
- Implemented heartbeat mechanism in WorkerActor to report status to JobCoordinatorGrain.
- Updated JobCoordinatorGrain to track agent telemetry and handle timeouts for inactive agents.
- Modified HashcatWrapper to parse telemetry data from Hashcat output, including GPU metrics.
- Adjusted job progress reporting to use AgentTelemetry instead of simple agent IDs.
- Improved error handling and cleanup processes in WorkerActor and JobCoordinatorGrain.
- Updated Blazor components to support new wordlist selection UI and validation.
- Changed data types in protobuf definitions for consistency with new telemetry metrics.

+++ DPCS/DPCS.Coordinator/Strategies/IJobStrategy.cs
@@ -12,6 +12,6 @@ public interface IJobStrategy
     Task<DictionaryWorkAssignment?> NextDictionaryChunkAsync(ulong hashRate);
     void CompleteChunk(string requestId);
     void FailChunk(string requestId);
-    double GetProgress();
+    float GetProgress();
     void HandleRecoveredPasswords(IEnumerable<RecoveredPassword> recoveredPasswords);
 }
